+++ main/weather_client/dwin_weather.c
@@ -132,8 +132,8 @@ void get_weather_handler(void* main_data, esp_event_base_t base, int32_t key, vo
     memset(description_WEATHER, 0, MAX_LEN_DESCRIPTION);
     strncpy(description_WEATHER, description[0], MAX_LEN_DESCRIPTION);
     for(int i=0; temp && temp[i]; i++){
-        temp_OUTDOOR[i] = atof(temp[i]);
-        temp_FEELS_LIKE[i] = atof(temp_feel[i]);
+        temp_OUTDOOR[i] = atoi(temp[i]);
+        temp_FEELS_LIKE[i] = atoi(temp_feel[i]);
         PoP[i] = atof(pop[i])*100;
     }
     xEventGroupSetBits(dwin_event_group, BIT_WEATHER_OK);
